Migrate telegram-bot page
